move compression functions to new module
